+++ accounting/src/apideck_accounting_unify/connections.py
@@ -1260,3 +1260,695 @@ class Connections(BaseSDK):
             http_res_text,
             http_res,
         )
+
+    def imports(
+        self,
+        *,
+        service_id: str,
+        unified_api: str,
+        consumer_id: Optional[str] = None,
+        app_id: Optional[str] = None,
+        credentials: Optional[
+            Union[models.Credentials, models.CredentialsTypedDict]
+        ] = None,
+        settings: OptionalNullable[
+            Union[models.Settings, models.SettingsTypedDict]
+        ] = UNSET,
+        metadata: OptionalNullable[Dict[str, Any]] = UNSET,
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+        http_headers: Optional[Mapping[str, str]] = None,
+    ) -> models.VaultConnectionsImportResponse:
+        r"""Import connection
+
+        Import an authorized connection.
+
+
+        :param service_id: Service ID of the resource to return
+        :param unified_api: Unified API
+        :param consumer_id: ID of the consumer which you want to get or push data from
+        :param app_id: The ID of your Unify application
+        :param credentials:
+        :param settings: Connection settings. Values will persist to `form_fields` with corresponding id
+        :param metadata: Attach your own consumer specific metadata
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        :param http_headers: Additional headers to set or replace on requests.
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
+
+        request = models.VaultConnectionsImportRequest(
+            consumer_id=consumer_id,
+            app_id=app_id,
+            service_id=service_id,
+            unified_api=unified_api,
+            connection_import_data=models.ConnectionImportData(
+                credentials=utils.get_pydantic_model(
+                    credentials, Optional[models.Credentials]
+                ),
+                settings=utils.get_pydantic_model(
+                    settings, OptionalNullable[models.Settings]
+                ),
+                metadata=metadata,
+            ),
+        )
+
+        req = self._build_request(
+            method="POST",
+            path="/vault/connections/{unified_api}/{service_id}/import",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=True,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            http_headers=http_headers,
+            _globals=models.VaultConnectionsImportGlobals(
+                consumer_id=self.sdk_configuration.globals.consumer_id,
+                app_id=self.sdk_configuration.globals.app_id,
+            ),
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.connection_import_data,
+                False,
+                False,
+                "json",
+                models.ConnectionImportData,
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["5XX"])
+
+        http_res = self.do_request(
+            hook_ctx=HookContext(
+                base_url=base_url or "",
+                operation_id="vault.connectionsImport",
+                oauth2_scopes=[],
+                security_source=get_security_from_env(
+                    self.sdk_configuration.security, models.Security
+                ),
+            ),
+            request=req,
+            error_status_codes=["400", "401", "402", "404", "422", "4XX", "5XX"],
+            retry_config=retry_config,
+        )
+
+        response_data: Any = None
+        if utils.match_response(http_res, "200", "application/json"):
+            return models.VaultConnectionsImportResponse(
+                create_connection_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.CreateConnectionResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+        if utils.match_response(http_res, "400", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.BadRequestResponseData
+            )
+            raise models.BadRequestResponse(data=response_data)
+        if utils.match_response(http_res, "401", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnauthorizedResponseData
+            )
+            raise models.UnauthorizedResponse(data=response_data)
+        if utils.match_response(http_res, "402", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.PaymentRequiredResponseData
+            )
+            raise models.PaymentRequiredResponse(data=response_data)
+        if utils.match_response(http_res, "404", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.NotFoundResponseData
+            )
+            raise models.NotFoundResponse(data=response_data)
+        if utils.match_response(http_res, "422", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnprocessableResponseData
+            )
+            raise models.UnprocessableResponse(data=response_data)
+        if utils.match_response(http_res, "4XX", "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "default", "application/json"):
+            return models.VaultConnectionsImportResponse(
+                unexpected_error_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.UnexpectedErrorResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = utils.stream_to_text(http_res)
+        raise models.APIError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    async def imports_async(
+        self,
+        *,
+        service_id: str,
+        unified_api: str,
+        consumer_id: Optional[str] = None,
+        app_id: Optional[str] = None,
+        credentials: Optional[
+            Union[models.Credentials, models.CredentialsTypedDict]
+        ] = None,
+        settings: OptionalNullable[
+            Union[models.Settings, models.SettingsTypedDict]
+        ] = UNSET,
+        metadata: OptionalNullable[Dict[str, Any]] = UNSET,
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+        http_headers: Optional[Mapping[str, str]] = None,
+    ) -> models.VaultConnectionsImportResponse:
+        r"""Import connection
+
+        Import an authorized connection.
+
+
+        :param service_id: Service ID of the resource to return
+        :param unified_api: Unified API
+        :param consumer_id: ID of the consumer which you want to get or push data from
+        :param app_id: The ID of your Unify application
+        :param credentials:
+        :param settings: Connection settings. Values will persist to `form_fields` with corresponding id
+        :param metadata: Attach your own consumer specific metadata
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        :param http_headers: Additional headers to set or replace on requests.
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
+
+        request = models.VaultConnectionsImportRequest(
+            consumer_id=consumer_id,
+            app_id=app_id,
+            service_id=service_id,
+            unified_api=unified_api,
+            connection_import_data=models.ConnectionImportData(
+                credentials=utils.get_pydantic_model(
+                    credentials, Optional[models.Credentials]
+                ),
+                settings=utils.get_pydantic_model(
+                    settings, OptionalNullable[models.Settings]
+                ),
+                metadata=metadata,
+            ),
+        )
+
+        req = self._build_request_async(
+            method="POST",
+            path="/vault/connections/{unified_api}/{service_id}/import",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=True,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            http_headers=http_headers,
+            _globals=models.VaultConnectionsImportGlobals(
+                consumer_id=self.sdk_configuration.globals.consumer_id,
+                app_id=self.sdk_configuration.globals.app_id,
+            ),
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.connection_import_data,
+                False,
+                False,
+                "json",
+                models.ConnectionImportData,
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["5XX"])
+
+        http_res = await self.do_request_async(
+            hook_ctx=HookContext(
+                base_url=base_url or "",
+                operation_id="vault.connectionsImport",
+                oauth2_scopes=[],
+                security_source=get_security_from_env(
+                    self.sdk_configuration.security, models.Security
+                ),
+            ),
+            request=req,
+            error_status_codes=["400", "401", "402", "404", "422", "4XX", "5XX"],
+            retry_config=retry_config,
+        )
+
+        response_data: Any = None
+        if utils.match_response(http_res, "200", "application/json"):
+            return models.VaultConnectionsImportResponse(
+                create_connection_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.CreateConnectionResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+        if utils.match_response(http_res, "400", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.BadRequestResponseData
+            )
+            raise models.BadRequestResponse(data=response_data)
+        if utils.match_response(http_res, "401", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnauthorizedResponseData
+            )
+            raise models.UnauthorizedResponse(data=response_data)
+        if utils.match_response(http_res, "402", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.PaymentRequiredResponseData
+            )
+            raise models.PaymentRequiredResponse(data=response_data)
+        if utils.match_response(http_res, "404", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.NotFoundResponseData
+            )
+            raise models.NotFoundResponse(data=response_data)
+        if utils.match_response(http_res, "422", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnprocessableResponseData
+            )
+            raise models.UnprocessableResponse(data=response_data)
+        if utils.match_response(http_res, "4XX", "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "default", "application/json"):
+            return models.VaultConnectionsImportResponse(
+                unexpected_error_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.UnexpectedErrorResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = await utils.stream_to_text_async(http_res)
+        raise models.APIError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    def token(
+        self,
+        *,
+        service_id: str,
+        unified_api: str,
+        consumer_id: Optional[str] = None,
+        app_id: Optional[str] = None,
+        request_body: Optional[
+            Union[
+                models.VaultConnectionsTokenRequestBody,
+                models.VaultConnectionsTokenRequestBodyTypedDict,
+            ]
+        ] = None,
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+        http_headers: Optional[Mapping[str, str]] = None,
+    ) -> models.VaultConnectionsTokenResponse:
+        r"""Authorize Access Token
+
+        Triggers exchanging persisted connection credentials for an access token and store it in Vault. Currently supported for connections with the `client_credentials` or `password` OAuth grant type.
+
+        Note:
+        - Do not include any credentials in the request body. This operation does not persist changes, but only triggers the exchange of persisted connection credentials for an access token.
+        - The access token will not be returned in the response. A 200 response code indicates the authorization was successful and that a valid access token was stored on the connection.
+        - The access token will be used for subsequent API requests.
+
+
+        :param service_id: Service ID of the resource to return
+        :param unified_api: Unified API
+        :param consumer_id: ID of the consumer which you want to get or push data from
+        :param app_id: The ID of your Unify application
+        :param request_body:
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        :param http_headers: Additional headers to set or replace on requests.
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
+
+        request = models.VaultConnectionsTokenRequest(
+            consumer_id=consumer_id,
+            app_id=app_id,
+            service_id=service_id,
+            unified_api=unified_api,
+            request_body=utils.get_pydantic_model(
+                request_body, Optional[models.VaultConnectionsTokenRequestBody]
+            ),
+        )
+
+        req = self._build_request(
+            method="POST",
+            path="/vault/connections/{unified_api}/{service_id}/token",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=False,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            http_headers=http_headers,
+            _globals=models.VaultConnectionsTokenGlobals(
+                consumer_id=self.sdk_configuration.globals.consumer_id,
+                app_id=self.sdk_configuration.globals.app_id,
+            ),
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.request_body,
+                False,
+                True,
+                "json",
+                Optional[models.VaultConnectionsTokenRequestBody],
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["5XX"])
+
+        http_res = self.do_request(
+            hook_ctx=HookContext(
+                base_url=base_url or "",
+                operation_id="vault.connectionsToken",
+                oauth2_scopes=[],
+                security_source=get_security_from_env(
+                    self.sdk_configuration.security, models.Security
+                ),
+            ),
+            request=req,
+            error_status_codes=["400", "401", "402", "404", "422", "4XX", "5XX"],
+            retry_config=retry_config,
+        )
+
+        response_data: Any = None
+        if utils.match_response(http_res, "200", "application/json"):
+            return models.VaultConnectionsTokenResponse(
+                get_connection_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.GetConnectionResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+        if utils.match_response(http_res, "400", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.BadRequestResponseData
+            )
+            raise models.BadRequestResponse(data=response_data)
+        if utils.match_response(http_res, "401", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnauthorizedResponseData
+            )
+            raise models.UnauthorizedResponse(data=response_data)
+        if utils.match_response(http_res, "402", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.PaymentRequiredResponseData
+            )
+            raise models.PaymentRequiredResponse(data=response_data)
+        if utils.match_response(http_res, "404", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.NotFoundResponseData
+            )
+            raise models.NotFoundResponse(data=response_data)
+        if utils.match_response(http_res, "422", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnprocessableResponseData
+            )
+            raise models.UnprocessableResponse(data=response_data)
+        if utils.match_response(http_res, "4XX", "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "default", "application/json"):
+            return models.VaultConnectionsTokenResponse(
+                unexpected_error_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.UnexpectedErrorResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = utils.stream_to_text(http_res)
+        raise models.APIError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    async def token_async(
+        self,
+        *,
+        service_id: str,
+        unified_api: str,
+        consumer_id: Optional[str] = None,
+        app_id: Optional[str] = None,
+        request_body: Optional[
+            Union[
+                models.VaultConnectionsTokenRequestBody,
+                models.VaultConnectionsTokenRequestBodyTypedDict,
+            ]
+        ] = None,
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+        http_headers: Optional[Mapping[str, str]] = None,
+    ) -> models.VaultConnectionsTokenResponse:
+        r"""Authorize Access Token
+
+        Triggers exchanging persisted connection credentials for an access token and store it in Vault. Currently supported for connections with the `client_credentials` or `password` OAuth grant type.
+
+        Note:
+        - Do not include any credentials in the request body. This operation does not persist changes, but only triggers the exchange of persisted connection credentials for an access token.
+        - The access token will not be returned in the response. A 200 response code indicates the authorization was successful and that a valid access token was stored on the connection.
+        - The access token will be used for subsequent API requests.
+
+
+        :param service_id: Service ID of the resource to return
+        :param unified_api: Unified API
+        :param consumer_id: ID of the consumer which you want to get or push data from
+        :param app_id: The ID of your Unify application
+        :param request_body:
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        :param http_headers: Additional headers to set or replace on requests.
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
+
+        request = models.VaultConnectionsTokenRequest(
+            consumer_id=consumer_id,
+            app_id=app_id,
+            service_id=service_id,
+            unified_api=unified_api,
+            request_body=utils.get_pydantic_model(
+                request_body, Optional[models.VaultConnectionsTokenRequestBody]
+            ),
+        )
+
+        req = self._build_request_async(
+            method="POST",
+            path="/vault/connections/{unified_api}/{service_id}/token",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=False,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            http_headers=http_headers,
+            _globals=models.VaultConnectionsTokenGlobals(
+                consumer_id=self.sdk_configuration.globals.consumer_id,
+                app_id=self.sdk_configuration.globals.app_id,
+            ),
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.request_body,
+                False,
+                True,
+                "json",
+                Optional[models.VaultConnectionsTokenRequestBody],
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["5XX"])
+
+        http_res = await self.do_request_async(
+            hook_ctx=HookContext(
+                base_url=base_url or "",
+                operation_id="vault.connectionsToken",
+                oauth2_scopes=[],
+                security_source=get_security_from_env(
+                    self.sdk_configuration.security, models.Security
+                ),
+            ),
+            request=req,
+            error_status_codes=["400", "401", "402", "404", "422", "4XX", "5XX"],
+            retry_config=retry_config,
+        )
+
+        response_data: Any = None
+        if utils.match_response(http_res, "200", "application/json"):
+            return models.VaultConnectionsTokenResponse(
+                get_connection_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.GetConnectionResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+        if utils.match_response(http_res, "400", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.BadRequestResponseData
+            )
+            raise models.BadRequestResponse(data=response_data)
+        if utils.match_response(http_res, "401", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnauthorizedResponseData
+            )
+            raise models.UnauthorizedResponse(data=response_data)
+        if utils.match_response(http_res, "402", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.PaymentRequiredResponseData
+            )
+            raise models.PaymentRequiredResponse(data=response_data)
+        if utils.match_response(http_res, "404", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.NotFoundResponseData
+            )
+            raise models.NotFoundResponse(data=response_data)
+        if utils.match_response(http_res, "422", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnprocessableResponseData
+            )
+            raise models.UnprocessableResponse(data=response_data)
+        if utils.match_response(http_res, "4XX", "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "default", "application/json"):
+            return models.VaultConnectionsTokenResponse(
+                unexpected_error_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.UnexpectedErrorResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = await utils.stream_to_text_async(http_res)
+        raise models.APIError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )