+++ accounting/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.9.3] accounting
 ### Releases
 - [PyPI v0.9.3] https://pypi.org/project/apideck-accounting-unify/0.9.3 - accounting
+
+## 2025-04-01 13:08:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.526.6 (2.563.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.10.0] accounting
+### Releases
+- [PyPI v0.10.0] https://pypi.org/project/apideck-accounting-unify/0.10.0 - accounting